requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
Appium-Python-Client==2.6.1
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.8.2
browserstack-local==1.2.2
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
colorama==0.4.4
cx-Oracle==7.0.0
DBUtils==1.4
dmPython==2.3
elasticsearch==7.15.2
et-xmlfile==1.1.0
exceptiongroup==1.1.0
execnet==2.0.2
Faker==6.6.3
h11==0.14.0
idna==2.8
importlib-metadata==3.7.3
iniconfig==1.1.1
JayDeBeApi==1.2.3
jdcal==1.4.1
JPype1==1.4.1
jsonpath==0.82
jsonschema==3.2.0
kafka==1.3.5
lxml==4.5.2
more-itertools==8.7.0
neo4j-driver==1.6.2
neotime==1.0.0
numpy==1.18.5
openpyxl==3.0.5
outcome==1.2.0
packaging==20.9
pandas==0.25.3
Pillow==7.0.0
pluggy==0.13.1
psutil==5.9.4
py==1.11.0
pycparser==2.21
pycrypto==2.6.1
pymongo==3.7.2
PyMySQL==0.9.3
pyparsing==2.4.7
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.4.4
pytest-html==1.19.0
pytest-instafail==0.4.2
pytest-metadata==2.0.4
pytest-multithreading-allure==1.0.8
pytest-parallel==0.1.1
pytest-rerunfailures==11.0
pytest-timeout==2.3.1
pytest-timestamps==0.1.4
pytest-xdist==3.5.0
python-dateutil==2.8.2
pytz==2019.3
PyYAML==5.4.1
redis==2.10.6
redis-py-cluster==1.3.6
requests==2.28.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2
tblib==1.7.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
wsproto==1.2.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0